Rename output to view.
