+++ javascripts/controller.js
@@ -4,7 +4,7 @@ let $ = require("jquery");
 let formatter = require("./formatter");
 let fbFactory = require("./fbFactory");
 let tmdb = require("./tmdb");
-let output = require("./output");
+let output = require("./view");
 let interactions = require("./interactions");
 
 // get value from users search and pass to ajax call
